saml-okta-main/requirements/prod.txt
# Everything we needed in production environment

Flask
Flask-Bootstrap
Flask-Login
pysaml2
Jinja2
requests